Parse imageset version from imageset name for OCM upsert
The current approach is to use metadata.addonImageSetVersion. However, the problem with
this is that the said field could be set to `latest`, which cause issues when posting
data to OCM, as `latest` is not recongized as a version. Instead, this PR parses the
version directly from the imageset name instead of relying on what is set in the metadata

Signed-off-by: Mayank Shah <m.shah@redhat.com>