Adding support for VERSION file
The AddonImageSet version assumes the version of the max
bundle in the main bundles directory.

Now, we are going to introduce support for Packages, then the
AddonImageSet version should optionally be detached from the bundles
version.

THis patch adds support for having a VERSION file at the root of the
addon directory. When the VERSION file exists, the version there will be
used, instead of probing the version from the bundles.

Signed-off-by: Amador Pahim <apahim@redhat.com>